integración del font con el back

+++ negocios/app/Http/Controllers/DefaultController.php
@@ -6,6 +6,8 @@ use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Arr;
+use Illuminate\Support\Collection;
 
 use App\Models\Intereses;
 use App\Models\Usuarios;
@@ -23,7 +25,8 @@ class DefaultController extends Controller {
 				   $oUsuarios = Usuarios::where('email', $email)->first();
 				   if(Hash::check($password, $oUsuarios->password)){
 					   if($oUsuarios->id_estado == 1){
-						return response()->json([ 'estado'=> true, 'resultado'=>  $oUsuarios ], 200);
+						$resultado = ['id'=> $oUsuarios->id, 'id_perfil'=> $oUsuarios->id_perfil, 'id_estado'=> $oUsuarios->id_estado, 'nombre'=> $oUsuarios->nombre, 'email'=> $oUsuarios->email];   
+						return response()->json([ 'estado'=> true, 'resultado'=>  $resultado ], 200);
 					   }
 					   return response()->json([ 'estado'=> false, 'mensaje'=> 'No tienes los permisos para acceder.'], 200);
 				   }
@@ -37,7 +40,7 @@ class DefaultController extends Controller {
 		}
 	}
 
-	public function obtenerUsuariosComunes(){
+	public function obtenerUsuarios(){
 		try {
 			$aResultado = [];
 			$oUsuarios = DB:: table('usuarios')
@@ -63,15 +66,23 @@ class DefaultController extends Controller {
 			return response()->json(['mensaje' => 'No se pudo comunicarse en el servidor.' ], 500);
 		}
 	}
-
-	public  function obtenerIntereses(){
+	public  function obtenerUsuario($id_usuario){
 		try {
-			if(count( Intereses::all() ) > 0){
-				return response()->json([ 'estado'=> true, 'resultado'=>  Intereses::all() ], 200);
+			if(Usuarios::where('id', $id_usuario)->exists()){
+				$oUsuarios = Usuarios::find($id_usuario, ['id', 'id_perfil', 'nombre', 'email']);
+				$oIntereses = DB::table('rel_usuarios_intereses')
+					->join('intereses', 'intereses.id', '=', 'rel_usuarios_intereses.id_interes')
+					->join('usuarios', 'usuarios.id', '=', 'rel_usuarios_intereses.id_usuario')
+					->select('intereses.id')
+					->where('usuarios.id', $id_usuario)
+					->orderBy('intereses.id', 'ASC')
+					->get();
+				$resultado  = ['id'=> $oUsuarios->id, 'id_perfil'=> $oUsuarios->id_perfil, 'nombre'=> $oUsuarios->nombre, 'email'=> $oUsuarios->email, 'intereses'=> Arr::pluck($oIntereses, 'id')];	
+                return response()->json([ 'estado'=> true, 'resultado'=> $resultado], 200);
 			}
-			return response()->json([ 'estado'=> false, 'mensaje' =>  'Sin datos registrados.' ], 200);
+			return response()->json([ 'estado'=> false, 'mensaje'=> 'No exite este usuario.'], 200);
 		} catch (\Throwable $th) {
-			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+			return response()->json(['mensaje' => 'No se pudo comunicarse en el servidor.' ], 500);
 		}
 	}
 	public function guardarUsuario(Request $request){
@@ -82,7 +93,7 @@ class DefaultController extends Controller {
 			$password  = $request->password;
 			$intereses = $request->intereses;
 
-			if($id_perfil && $nombre && $email && $password && $intereses){
+			if($id_perfil && $nombre && $email && $password){
 
 				if(Usuarios::where('email', $email)->exists()){
 					return response()->json([ 'estado'=> false, 'mensaje'=> 'Email registrado intentelo con otro.'], 200);
@@ -95,12 +106,13 @@ class DefaultController extends Controller {
 				$oUsuarios->email     = $email;
 				$oUsuarios->password  = Hash::make($password);
 				$oUsuarios->save();
-
-				foreach($intereses as $interes){
-					$oRelUsuariosIntereses = new RelUsuariosIntereses;
-					$oRelUsuariosIntereses->id_usuario = $oUsuarios->id;
-					$oRelUsuariosIntereses->id_interes = $interes;
-					$oRelUsuariosIntereses->save();
+				if(!empty($intereses)){
+					foreach($intereses as $interes){
+						$oRelUsuariosIntereses = new RelUsuariosIntereses;
+						$oRelUsuariosIntereses->id_usuario = $oUsuarios->id;
+						$oRelUsuariosIntereses->id_interes = $interes;
+						$oRelUsuariosIntereses->save();
+					}
 				}
 				return response()->json([ 'estado'=> true, 'mensaje'=> 'Datos guardado correctamente.'], 200 );
 			}
@@ -112,12 +124,11 @@ class DefaultController extends Controller {
 	public function modificarUsuario(Request $request, $id_usuario){
 		try {
 				$id_perfil = (int)$request->id_perfil;
-				$id_estado = (int)$request->id_estado;
 				$nombre    = $request->nombre;
 				$password  = $request->password;
 				$intereses = $request->intereses;
 
-				if($nombre && $intereses){
+				if($nombre){
 					$oUsuarios = Usuarios::find($id_usuario);
 					if($oUsuarios){
 						$oUsuarios->nombre = $nombre;
@@ -125,23 +136,22 @@ class DefaultController extends Controller {
 						if($id_perfil){
 							$oUsuarios->id_perfil = $id_perfil;
 						}
-						if($id_estado){
-							$oUsuarios->id_estado = $id_estado;
-						}
 						if($password){
 							$oUsuarios->password  = Hash::make($password);
 						}
 						$oUsuarios->save();
 
-						$oRelUsuariosIntereses =  RelUsuariosIntereses::where('id_usuario', $id_usuario)->get();
-						foreach($oRelUsuariosIntereses as $item){
-							$item->delete();
-						}
-						foreach($intereses as $interes){
-							$oRelUsuariosIntereses = new RelUsuariosIntereses;
-							$oRelUsuariosIntereses->id_usuario = $id_usuario;
-							$oRelUsuariosIntereses->id_interes = $interes;
-							$oRelUsuariosIntereses->save();
+						if(!empty($intereses)){
+							$oRelUsuariosIntereses =  RelUsuariosIntereses::where('id_usuario', $id_usuario)->get();
+							foreach($oRelUsuariosIntereses as $item){
+								$item->delete();
+							}
+							foreach($intereses as $interes){
+								$oRelUsuariosIntereses = new RelUsuariosIntereses;
+								$oRelUsuariosIntereses->id_usuario = $id_usuario;
+								$oRelUsuariosIntereses->id_interes = $interes;
+								$oRelUsuariosIntereses->save();
+							}
 						}
 						return response()->json([ 'estado'=> true, 'mensaje'=> 'Datos modificados correctamente.'], 200 );
 					}
@@ -167,5 +177,99 @@ class DefaultController extends Controller {
 		} catch (\Throwable $th) {
 			return response()->json(['mensaje' => 'No se pudo eliminar el usuario.' ], 500);
 		}
-	}  
+	}
+	public function bloquearUsuario($id_usuario){
+		try {
+			if(Usuarios::where('id', $id_usuario)->exists()){
+                $oUsuarios = Usuarios::find($id_usuario);
+				$oUsuarios->id_estado = 2;
+				$oUsuarios->save();
+
+				return response()->json([ 'estado'=> true, 'mensaje'=> 'El usuario fue bloqueado correctamente.'], 200 );
+			}
+			return response()->json([ 'estado'=> false, 'mensaje'=> 'No exite este usuario.'], 200);
+		} catch (\Throwable $th) {
+			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+		}
+	}
+	public function ActivoUsuario($id_usuario){
+		try {
+			if(Usuarios::where('id', $id_usuario)->exists()){
+                $oUsuarios = Usuarios::find($id_usuario);
+				$oUsuarios->id_estado = 1;
+				$oUsuarios->save();
+
+				return response()->json([ 'estado'=> true, 'mensaje'=> 'El usuario fue activado correctamente.'], 200 );
+			}
+			return response()->json([ 'estado'=> false, 'mensaje'=> 'No exite este usuario.'], 200);
+		} catch (\Throwable $th) {
+			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+		}
+	}
+	public  function obtenerIntereses(){
+		try {
+			if(count( Intereses::all() ) > 0){
+				return response()->json([ 'estado'=> true, 'resultado'=>  Intereses::all(['id as value', 'nombre as label']) ], 200);
+			}
+			return response()->json([ 'estado'=> false, 'mensaje' =>  'Sin datos registrados.' ], 200);
+		} catch (\Throwable $th) {
+			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+		}
+	}
+	public  function guardarInteresesUsuario(Request $request){
+		try {
+			$id_usuario   = $request->id;
+			$intereses    = $request->intereses;
+			if($id_usuario && $intereses){
+				if(Usuarios::where('id', $id_usuario)->exists()){
+					$oRelUsuariosIntereses =  RelUsuariosIntereses::where('id_usuario', $id_usuario)->get();
+					foreach($oRelUsuariosIntereses as $item){
+						$item->delete();
+					}
+					foreach($intereses as $interes){
+						$oRelUsuariosIntereses = new RelUsuariosIntereses;
+						$oRelUsuariosIntereses->id_usuario = $id_usuario;
+						$oRelUsuariosIntereses->id_interes = $interes;
+						$oRelUsuariosIntereses->save();
+					}
+                    return response()->json([ 'estado'=> true, 'mensaje'=> 'Se guardo sus intereses, se actualizara la lista de usuario en común en unos 30 segundos.'], 200 ); 
+				}
+				return response()->json([ 'estado'=> false, 'mensaje'=> 'No exite este usuario.'], 200);
+			}
+			return response()->json([ 'estado'=> false, 'mensaje' =>  'Inconsistencia de datos.' ], 200);
+		} catch (\Throwable $th) {
+			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+		}
+	}
+	public function obtenerInteresComunes(Request $request){
+		try {
+			$resultado = [];
+			$id_usuario   = $request->id;
+			$intereses    = $request->intereses;
+			if($id_usuario && $intereses){
+               $oUsuarios = DB::table('usuarios')->select('id','nombre')->whereNotIn('id', [$id_usuario])->get();				
+			   foreach($oUsuarios as $usuario){
+					$oIntereses = DB::table('rel_usuarios_intereses')
+						->join('intereses', 'intereses.id', '=', 'rel_usuarios_intereses.id_interes')
+						->join('usuarios', 'usuarios.id', '=', 'rel_usuarios_intereses.id_usuario')
+						->select('intereses.nombre')
+						->where('usuarios.id', $usuario->id)
+						->whereIn('intereses.id', $intereses)
+						->get();
+					if(count($oIntereses) > 1){
+						$aIntereses = [];
+                       foreach($oIntereses as $interes){
+                          $aIntereses[] = $interes->nombre;
+					   }
+					   $resultado[] = ['usuario'=> $usuario->nombre, 'intereses' => $aIntereses];
+					}
+			   }
+			   return response()->json([ 'estado'=> true, 'resultado' =>  $resultado ], 200);
+			}
+			return response()->json([ 'estado'=> false, 'mensaje' =>  'Inconsistencia de datos.' ], 200);
+		} catch (\Throwable $th) {
+			return response()->json(['mensaje' => 'No se pudo obtener los datos.' ], 500);
+		}
+
+	}   
 }
